push push

+++ srcs/ft_printf_utils_two.c
@@ -6,7 +6,7 @@
 /*   By: ajearuth <ajearuth@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/29 13:34:31 by ajearuth          #+#    #+#             */
-/*   Updated: 2021/08/29 15:51:10 by ajearuth         ###   ########.fr       */
+/*   Updated: 2021/08/30 09:19:35 by ajearuth         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@ int	ft_print_p(va_list arg)
 
 	adress = va_arg(arg, void *);
 	ft_putstr_fd("0x", 1);
-	ft_putnbr_base_fd((unsigned long long int)adress, "0123456789abcdef", 1);
-    return(ft_nbrlen((unsigned long long int)adress, 16) + 2);
+	ft_putnbr_base_fd((int)adress, "0123456789abcdef", 1);
+    return(ft_nbrlen((unsigned int)adress, 16) + 2);
 }
 
 int	ft_print_x(va_list arg)
@@ -40,7 +40,7 @@ int	ft_print_upp_x(va_list arg)
 	return(ft_nbrlen(nb, 16));
 }
 
-int ft_nbrlen(unsigned long long int nb, int base)
+int ft_nbrlen(unsigned int nb, int base)
 {
     unsigned int  i;
 
@@ -53,7 +53,7 @@ int ft_nbrlen(unsigned long long int nb, int base)
     return(i);
 }
 
-void	ft_putnbr_uint_fd(unsigned long long int n, int fd)
+void	ft_putnbr_uint_fd(unsigned int n, int fd)
 {
 	if (n > 9)
 	{
